weather3/requirements.txt
geopy==2.2.0
requests==2.26.0
pytz==2021.3
Pillow==9.0.1
h3==3.7.3
timezonefinder==5.2.0
